backend/requirements.txt
# Core Flask
Flask==2.3.2
Flask-CORS==4.0.0
Werkzeug==2.3.6


psycopg2-binary==2.9.9  
SQLAlchemy==2.0.23   


# Machine Learning
scikit-learn==1.3.2
joblib==1.3.1

# Security & Encryption
cryptography==41.0.0
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Rate Limiting
Flask-Limiter==3.5.0

# PDF Generation
reportlab==4.0.4
Pillow==10.0.0

# Development (optional)
pytest==7.4.0
python-decouple==3.8
